* Adjustment: Initial work on moving x86Unix platform code into a more common POSIX platform.
