+++ Engine/source/platformMac/macMain.mm
@@ -23,14 +23,18 @@
 #import "app/mainLoop.h"
 #import "platform/platformInput.h"
 #import "console/console.h"
+#import "platformPOSIX/POSIXStdConsole.h"
+#import "platformPOSIX/POSIXState.h"
 
 extern void InitWindowingSystem();
 
 //------------------------------------------------------------------------------
 void Platform::init()
 {
-   
-   Con::printf("Initializing platform...");
+    StdConsole::create();
+    stdConsole->enable(true);
+
+    Con::printf("Initializing platform...");
    
    // Set the platform variable for the scripts
    Con::setVariable( "$platform", "MacOSX" );
