+++ Engine/source/platformPOSIX/POSIXStrings.cpp
