+++ Engine/source/platformPOSIX/POSIXTime.cpp
@@ -20,6 +20,8 @@
 // IN THE SOFTWARE.
 //-----------------------------------------------------------------------------
 
+#ifndef __APPLE__
+
 #include "platformX86UNIX/platformX86UNIX.h"
 #include "platform/platformTimer.h"
 #include "time.h"
@@ -158,4 +160,4 @@ void Platform::sleep(U32 ms)
 	// note: this will overflow if you want to sleep for more than 49 days. just so ye know.
 	usleep( ms * 1000 );
 }
-	    
+#endif
